pyproject.toml
[project]
name = "rag-file-monitor"
version = "0.1.0"
description = "File monitoring tool for RAG system with Qdrant vector database"
readme = "README.md"
requires-python = ">=3.8,<3.13"
dependencies = [
    "watchdog>=3.0.0",
    "torch>=2.0.0",
    "qdrant-client>=1.1.0",
    "sentence-transformers>=2.2.0",
    "numpy>=1.21.0,<2.0.0",
    "PyPDF2>=3.0.0",
    "python-docx>=1.1.0",
    "beautifulsoup4>=4.12.0",
    "pyyaml>=6.0",
    "python-magic>=0.4.27",
    "chardet>=5.2.0",
    "click>=8.1.0",
    "tqdm>=4.66.0",
    "libmagic>=0.4",
    "psutil>=7.0.0"
]

[project.scripts]
rag-monitor = "rag_file_monitor.file_monitor:main"
rag-manage = "rag_file_monitor.manage_deleted:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["rag_file_monitor"]